Fixup for Python 3.10
Python 3.8 deprecated the use of floats when an integer argument is required for builtin and extension functions. With Python 3.10, this is now an error.

Commit inspired by https://github.com/Electron-Cash/Electron-Cash/commit/a55dadc51075a8385e3d3c2c065fd0b92be107bc  with minor changes to try to avoid the int conversion when float cast can be avoided upstream.

+++ electroncash_gui/qt/qrreader/video_overlay.py
@@ -67,7 +67,7 @@ class QrReaderVideoOverlay(QtWidgets.QWidget):
         self.bg_rect_pen = QPen()
         self.bg_rect_pen.setColor(Qt.black)
         self.bg_rect_pen.setStyle(Qt.DotLine)
-        self.bg_rect_fill = QColor(255, 255, 255, 255 * self.BG_RECT_OPACITY)
+        self.bg_rect_fill = QColor(255, 255, 255, int(255 * self.BG_RECT_OPACITY))
 
         self.qr_finder = QSvgRenderer(":icons/qr_finder.svg")
